Actuary/requirements.txt
asgiref==3.2.7
brotlipy==0.7.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.3
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
numpy==1.18.5
pandas==1.0.4
pycparser==2.20
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
win-inet-pton==1.1.0
wincertstore==0.2
django-rest-auth==0.9.5
django-allauth==0.42.0
gunicorn==19.7.1
whitenoise==3.3.1
